cc26x2: ccfg registers are u32

+++ chips/cc26x2/src/ccfg.rs
@@ -99,7 +99,7 @@ impl Registers {
 }
 
 register_bitfields![
-    u8,
+    u32,
     ExtLfClk [
         // Unsigned value pin selection
         DIO OFFSET(24) NUMBITS(8) [],
